+++ server/lib/v1/ILecture.ts
